only allow authorized users to call main endpoints

+++ inst/plumber.R
@@ -14,8 +14,12 @@
 function (repourl = "", repo, issue_id) {
 
     if (nchar (repourl) == 0L) {
-          return ("Error: Issue template has no 'repourl'")
-      }
+        return ("Error: Issue template has no 'repourl'")
+    }
+    user <- get_github_user ()
+    if (!user %in% authorized_users) {
+        return ("Only authorized users may call this endpoint")
+    }
 
     repourl <- as.character (repourl) [1]
     repo <- as.character (repo) [1]
@@ -23,8 +27,8 @@ function (repourl = "", repo, issue_id) {
 
     template_chk <- roreviewapi::check_issue_template (repo, issue_id)
     if (!attr (template_chk, "proceed_with_checks")) {
-          return (template_chk)
-      }
+        return (template_chk)
+    }
 
     logfiles <- roreviewapi::stdout_stderr_cache (repourl)
 
@@ -62,8 +66,12 @@ function (repourl = "", repo, issue_id) {
 function (repourl = "") {
 
     if (nchar (repourl) == 0L) {
-          return ("Error: Issue template has no 'repourl'")
-      }
+        return ("Error: Issue template has no 'repourl'")
+    }
+    user <- get_github_user ()
+    if (!user %in% authorized_users) {
+        return ("Only authorized users may call this endpoint")
+    }
 
     out <- roreviewapi::editor_check (repourl,
         post_to_issue = FALSE
@@ -92,8 +100,8 @@ function (n = 10) {
 function (repo = "ropensci/software-review", issue_num) {
 
     if (!is.integer (issue_num) & length (issue_num) != 1L) {
-          return (NULL)
-      }
+        return (NULL)
+    }
 
     roreviewapi::stats_badge (repo, issue_num)
 }
@@ -128,8 +136,10 @@ function (n = 10) {
 #* @get /clear_cache
 function () {
 
-    cache_dir <- Sys.getenv ("PKGCHECK_CACHE_DIR")
-    chk <- unlink (cache_dir, recursive = TRUE)
+    cache_dir <- options ("pkgcheck.cache_dir")
+    if (dir.exists (cache_dir)) {
+        chk <- unlink (cache_dir, recursive = TRUE)
+    }
 
     ifelse (chk == 0,
         "Cache directory successfully removed",
